refactor: アカウント作成画面のファイルを整理した
