+++ client/Hi/Hi/View/AccountCreate/AccountCreateView.swift
@@ -1,0 +1,56 @@
+//
+//  AccountCreateView.swift
+//  Hi
+//
+//  Created by Yuma on 2024/04/11.
+//
+
+import SwiftUI
+
+struct AccountCreateNameView: View {
+    @EnvironmentObject var router: NavigationRouter
+    
+    var body: some View {
+        AccountSettingName(nextButtonLabel: "次へ", isShowBackButton: false, action: {
+            router.navigateToView(destination: .accountCreateUserID)
+        })
+        .navigationBarHidden(true)
+    }
+}
+
+struct AccountCreateUserID: View {
+    @EnvironmentObject var router: NavigationRouter
+    
+    var body: some View {
+        AccountSettingUserID(nextButtonLabel: "次へ", action: {
+            router.navigateToView(destination: .accountCreateWakeUpTime)
+        })
+        .navigationBarHidden(true)
+    }
+}
+
+struct AccountCreateWakeUpTimeView: View {
+    @EnvironmentObject var router: NavigationRouter
+    
+    var body: some View {
+        AccountSettingWakeUpTime(nextButtonLabel: "次へ", action: {
+            router.navigateToView(destination: .accountCreateDayOfWeek)
+        })
+        .navigationBarHidden(true)
+    }
+}
+
+struct AccountCreateDayOfWeekView: View {
+    @EnvironmentObject var router: NavigationRouter
+    
+    var body: some View {
+        AccountSettingDayOfWeek(nextButtonLabel: "Hiを始める！", action: {
+            router.navigateToView(destination: .main)
+        })
+        .navigationBarHidden(true)
+    }
+}
+
+#Preview {
+    AccountCreateNameView()
+}
